[CRYPMB-371] - fixed deposit bug, added PaymentStatus component

Approved-by: Valery Zakharchenko <vzakharchenko@newagesol.com>
Approved-by: Vitaliy Koshovyi <vkoshovyi@newagesol.com>
Approved-by: Eugene Husak <ehusak@newagesol.com>

+++ src/graphql/mutations/payment.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 const addPaymentMutation = gql`mutation createClientPayment(
-  $amount: Int!,
+  $amount: Float!,
   $currency: String!
   $paymentType: String!,
   $externalReference: String,
